+++ src/main/java/task/service/DepartmentService.java
@@ -17,7 +17,7 @@ public interface DepartmentService {
 
     Double averageSalary(String departmentName);
 
-    Integer getDepartmentLecturersCount(String departmentName);
+    Long getDepartmentLecturersCount(String departmentName);
 
     List<Department> getAll();
 }
